// notelab-client/src/connectors/gmail/src/types.ts
export type GmailMessagePartHeader = {
  name: string;
  value: string;
};

export type GmailMessagePartBody = {
  attachmentId?: string;
  data?: string;
  size?: number;
};

export type GmailMessagePart = {
  body?: GmailMessagePartBody;
  filename?: string;
  headers?: GmailMessagePartHeader[];
  mimeType?: string;
  partId?: string;
  parts?: GmailMessagePart[];
};

export type GmailMessage = {
  historyId?: string;
  id: string;
  internalDate?: string;
  labelIds?: string[];
  payload?: GmailMessagePart;
  raw?: string;
  sizeEstimate?: number;
  snippet?: string;
  threadId: string;
};

export type GmailDraft = {
  id: string;
  message?: GmailMessage;
};

export type GmailListDraftsResponse = {
  drafts?: GmailDraft[];
  nextPageToken?: string;
  resultSizeEstimate?: number;
};

export type GmailMessageAttachment = {
  attachmentId: string;
  data?: string;
  size?: number;
};

export type GmailAttachmentSummary = {
  attachmentId?: string;
  filename?: string;
  mimeType?: string;
  partId?: string;
  size?: number;
};
// ...
export type GmailMessageSummary = {
  attachments?: GmailAttachmentSummary[];
  bcc?: string;
  bodyHtml?: string;
  bodyText?: string;
  cc?: string;
  date?: string;
  from?: string;
  historyId?: string;
  id: string;
  internalDate?: string;
  labelIds?: string[];
  sizeEstimate?: number;
  snippet?: string;
  subject?: string;
  senderPhotoUrl?: string;
  threadId: string;
  to?: string;
};